refactor: extract bash wrapper classification
